requirements.txt
Django~=5.0.7
djangorestframework~=3.15.2
typing_extensions~=4.12.2
pillow~=10.4.0
pip~=24.2
attrs~=23.2.0
DateTime~=5.5
asgiref~=3.8.1
sqlparse~=0.5.1
psycopg2~=2.9.9
colorama~=0.4.6
PyYAML~=6.0.2
requests~=2.32.3
click~=8.1.7
uritemplate~=4.1.1
inflection~=0.5.1
pytz~=2024.1
python-dotenv~=1.0.1